pyproject.toml
[project]
name = "trading-knowledge-base"
version = "1.0.0"
description = "AI-powered quantitative finance research system with crew-based orchestration and memory management"
authors = [{ name = "Trading Knowledge Base Team", email = "team@tradingkb.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.150.0,<1.0.0",
    "requests",
    "google-generativeai",
    "sentence-transformers",
    "faiss-cpu",
    "psycopg2-binary",
    "redis",
    "python-dotenv"
]

[project.scripts]
trading-kb = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"

[project.urls]
Homepage = "https://github.com/yourusername/trading-knowledge-base"
Documentation = "https://github.com/yourusername/trading-knowledge-base#readme"
Repository = "https://github.com/yourusername/trading-knowledge-base.git"
